+++ raiden-dapp/src/store/notifications/types.d.ts
@@ -1,11 +1,14 @@
 import { NotificationImportance } from '@/store/notifications/notification-importance';
 import { NotificationContext } from '@/store/notifications/notification-context';
+import { RouteNames } from '@/router/route-names';
 
 export interface NotificationPayload {
   readonly id: number;
+  readonly icon?: string;
   readonly title: string;
   readonly description: string;
-  readonly icon?: string;
+  readonly link?: string;
+  readonly dappRoute?: RouteNames;
   readonly importance: NotificationImportance;
   readonly context: NotificationContext;
   readonly duration: number;