+++ raiden-dapp/src/views/TransferRoute.vue
@@ -20,7 +20,7 @@
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
-import { mapGetters } from 'vuex';
+import { mapState, mapGetters } from 'vuex';
 import TransferHeaders from '@/components/transfer/TransferHeaders.vue';
 import TransferInputs from '@/components/transfer/TransferInputs.vue';
 import TransactionList from '@/components/transaction-history/TransactionList.vue';
@@ -29,6 +29,12 @@ import NoChannelsDialog from '@/components/dialogs/NoChannelsDialog.vue';
 import { RaidenChannel } from 'raiden-ts';
 import { BigNumber, constants } from 'ethers';
 import { Token, TokenModel } from '@/model/types';
+import { NotificationPayload } from '../store/notifications/types';
+import { NotificationImportance } from '../store/notifications/notification-importance';
+import { NotificationContext } from '../store/notifications/notification-context';
+import { RouteNames } from '@/router/route-names';
+
+const ONE_DAY = new Date(0).setUTCHours(24);
 
 @Component({
   components: {
@@ -39,14 +45,45 @@ import { Token, TokenModel } from '@/model/types';
     NoChannelsDialog,
   },
   computed: {
+    ...mapState(['stateBackupReminderDateMs']),
     ...mapGetters(['tokens', 'channelWithBiggestCapacity', 'openChannels']),
   },
 })
 export default class TransferRoute extends Vue {
+  stateBackupReminderDateMs!: number;
   tokens!: TokenModel[];
   channelWithBiggestCapacity!: (
     tokenAddress: string
   ) => RaidenChannel | undefined;
+
+  mounted() {
+    const currentTime = new Date().getTime();
+
+    if (
+      this.stateBackupReminderDateMs === 0 ||
+      currentTime > this.stateBackupReminderDateMs + ONE_DAY
+    ) {
+      this.pushStateBackupNotification(currentTime);
+    }
+  }
+
+  pushStateBackupNotification(currentTime: number): void {
+    const stateBackupReminder = {
+      icon: this.$t('notifications.backup-state.icon') as string,
+      title: this.$t('notifications.backup-state.title') as string,
+      link: this.$t('notifications.backup-state.link') as string,
+      dappRoute: RouteNames.ACCOUNT_BACKUP,
+      description: this.$t('notifications.backup-state.description') as string,
+      importance: NotificationImportance.HIGH,
+      context: NotificationContext.WARNING,
+    } as NotificationPayload;
+
+    this.$store.commit('updateStateBackupReminderDate', currentTime);
+    this.$store.commit(
+      'notifications/notificationAddOrReplace',
+      stateBackupReminder
+    );
+  }
 
   get noTokens(): boolean {
     return this.tokens.length === 0;