+++ raiden-dapp/tests/unit/data/mock-data.ts
@@ -174,6 +174,8 @@ export class TestData {
     title: 'Channel Settlement',
     description: 'Channel with 0x09123456789 was settled.',
     icon: 'notification_settle',
+    link: 'Visit the Withdrawal menu',
+    dappRoute: RouteNames.ACCOUNT_WITHDRAWAL,
     display: false,
     duration: 5000,
     importance: NotificationImportance.HIGH,